requirements.txt
fastapi
uvicorn
pillow
python-multipart
onnxruntime
numpy
